// services/ms-users/src/users/interfaces/user.interface.ts
export interface User {
    id: string;
    name: string;
    birthdate: string;
    email: string;
    phone_number: string;
    city: string;
    state: string;
    country: string;
    address: string;
    complement: string;
    zip_code: string;
    created_at: string;
    updated_at: string;
}